+++ src/ArarasHealthHub.Application/Features/Orders/Dtos/OrderItemDto.cs
@@ -8,11 +8,11 @@ namespace ArarasHealthHub.Application.Features.Orders.Dtos
     public class OrderItemDto
     {
         public int Id { get; set; }
-        public int AvailableQuantity { get; set; }
-        public int RequestedQuantity { get; set; }
-        public int ApprovedQuantity { get; set; }
-        public int ReservedQuantity { get; set; }
-        public int ActualQuantity { get; set; }
+        public decimal AvailableQuantity { get; set; }
+        public decimal RequestedQuantity { get; set; }
+        public decimal ApprovedQuantity { get; set; }
+        public decimal ReservedQuantity { get; set; }
+        public decimal ActualQuantity { get; set; }
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
     }
